feat(box-config) default to two workers
Provide a default for vb.cpus that enables Kong to start with multiple
workers. This more closely mimics a Kong node running in production
and increases the chances for developers to discover concurrency bugs
before issuing a PR.

Default is set to two under the assumption that most developers have
a multi-cpu machine these days. VirtualBox will silently start with
a value of one, effectively reverting to the behavior we have today.

Tested on Mac only.